requirements.txt
Django==3.1.2
django-cors-headers==3.5.0
django-filter==2.4.0
djangorestframework==3.12.1
gunicorn==20.0.4
pip-tools==5.3.1
pytz==2020.1
sqlparse==0.3.1
